% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_SS_inputs.R
\name{copy_SS_inputs}
\alias{copy_SS_inputs}
\title{Copy a the Stock Synthesis input files from one directory to another}
\usage{
copy_SS_inputs(dir.old = NULL, dir.new = NULL, create.dir.new = TRUE,
  overwrite = FALSE, use_ss_new = FALSE, verbose = TRUE)
}
\arguments{
\item{dir.old}{Location of model files to be copied.}

\item{dir.new}{New location to which the files should be copied.}

\item{create.dir.new}{Create dir.new directory if it doesn't exist already?}

\item{overwrite}{Overwrite existing files with matching names?}

\item{use_ss_new}{Use .ss_new files instead of original inputs?}

\item{verbose}{Return updates of function progress to the R console?}
}
\value{
Doesn't return anything
}
\description{
Reads the starter.ss file to figure out the names of the control and
data files, than copies those files along with starter.ss, forecast.ss,
and wtatage.ss (if present) to a new directory, as specified.
}
\examples{

  \dontrun{
    copy_SS_inputs(dir.old='c:/SS/old_model',
                   dir.new='c:/SS/new_model')
  }

}
\author{
Ian Taylor
}
