% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_SS_models_parallel.R
\name{run_models_parallel_base}
\alias{run_models_parallel_base}
\title{Base function to run a Stock Synthesis model.}
\usage{
run_models_parallel_base(dir, exe, exe_in_path, skipfinished, model,
  extras, OS, systemcmd, intern)
}
\arguments{
\item{dir}{A director List of directories containing the model files}

\item{exe}{The path to the exe}

\item{exe_in_path}{logical. If TRUE, will look for exe in the PATH. If FALSE,
will look for exe in the model folders. Default = FALSE.}

\item{skipfinished}{Skip any folders that already contain a Report.sso file.
This can be helpful if the function is interrupted.}

\item{model}{Executable name}

\item{extras}{Additional commands to use when running SS. Default = "-nox"
will reduce the amount of command-line output.}

\item{systemcmd}{Should R call SS using "system" function instead of "shell".
This may be required when running R in Emacs. Default = FALSE.}

\item{intern}{Show output in the R console or save to a file?
Default = TRUE.}
}
\value{
Returns table showing which directories had model run and which
had errors like missing executable or Report.sso already present
}
\description{
Runs SS in a directory
}
\seealso{
\code{\link{run_SS_models_parallel}}
}
\author{
Ian Taylor, Kathryn Doering
}
